refactor: 사용하지 않는 의존성 제거 및 오타 수정, 클래스 위치 변경
