+++ src/main/java/com/orderingsystem/common/exception/DomainException.java
@@ -1,4 +1,4 @@
-package com.orderingsystem.common.exceptioin;
+package com.orderingsystem.common.exception;
 
 public class DomainException extends RuntimeException {
     public DomainException(String message) {
