+++ src/main/java/com/orderingsystem/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.orderingsystem.common.exceptioin;
+package com.orderingsystem.common.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
