+++ src/main/java/com/orderingsystem/customer/domain/model/Customer.java
@@ -2,8 +2,6 @@ package com.orderingsystem.customer.domain.model;
 
 import com.orderingsystem.common.domain.AggregateRoot;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import java.util.UUID;
