+++ src/main/java/com/orderingsystem/order/application/publisher/OrderCancelledPaymentRequestMessagePublisher.java
@@ -2,7 +2,6 @@ package com.orderingsystem.order.application.publisher;
 
 import com.orderingsystem.common.domain.publisher.DomainEventPublisher;
 import com.orderingsystem.order.domain.event.OrderCancelledEvent;
-import com.orderingsystem.order.domain.event.OrderCreateEvent;
 
 public interface OrderCancelledPaymentRequestMessagePublisher extends DomainEventPublisher<OrderCancelledEvent> {
 }
