+++ src/main/java/com/orderingsystem/order/domain/event/OrderEvent.java
@@ -3,7 +3,9 @@ package com.orderingsystem.order.domain.event;
 import com.orderingsystem.common.domain.event.DomainEvent;
 import com.orderingsystem.order.domain.model.Order;
 import java.time.ZonedDateTime;
+import lombok.Getter;
 
+@Getter
 public abstract class OrderEvent implements DomainEvent<Order> {
 
     private final Order order;
@@ -12,14 +14,6 @@ public abstract class OrderEvent implements DomainEvent<Order> {
     public OrderEvent(Order order, ZonedDateTime createdAt) {
         this.order = order;
         this.createdAt = createdAt;
-    }
-
-    public Order getOrder() {
-        return order;
-    }
-
-    public ZonedDateTime getCreatedAt() {
-        return createdAt;
     }
 
 }
