+++ src/main/java/com/orderingsystem/order/domain/exception/OrderDomainException.java
@@ -1,6 +1,6 @@
 package com.orderingsystem.order.domain.exception;
 
-import com.orderingsystem.common.exceptioin.DomainException;
+import com.orderingsystem.common.exception.DomainException;
 
 public class OrderDomainException extends DomainException {
     public OrderDomainException(String message) {
