+++ src/main/java/com/orderingsystem/payment/application/publisher/PaymentCancelledMessagePublisher.java
@@ -2,7 +2,6 @@ package com.orderingsystem.payment.application.publisher;
 
 import com.orderingsystem.common.domain.publisher.DomainEventPublisher;
 import com.orderingsystem.payment.domain.event.PaymentCancelledEvent;
-import com.orderingsystem.payment.domain.event.PaymentCompletedEvent;
 
 public interface PaymentCancelledMessagePublisher extends DomainEventPublisher<PaymentCancelledEvent> {
 }
