+++ src/main/java/com/orderingsystem/restaurant/domain/exception/RestaurantNotFoundException.java
@@ -1,6 +1,6 @@
 package com.orderingsystem.restaurant.domain.exception;
 
-import com.orderingsystem.common.exceptioin.DomainException;
+import com.orderingsystem.common.exception.DomainException;
 
 public class RestaurantNotFoundException extends DomainException {
     public RestaurantNotFoundException(String message) {
